+++ src/main/java/com/flamingos/osp/bean/OspProfAcademicsBean.java
@@ -9,106 +9,147 @@ public class OspProfAcademicsBean implements Serializable {
 
   private static final Logger logger = Logger.getLogger(OspProfAcademicsBean.class);
   private static final long serialVersionUID = 1L;
-  private int profAcdmcId;
+  private Long profAcdmcId;
   private String profAcdmcName;
   private String profAcdmcDesc;
   private String profAcdmcUniversity;
   private String profAcdmcPassYear;
-  private long profId;
-  private int activeStatus;
+  private Long profId;
+  private Integer activeStatus;
   private Date createdTs;
   private Date updatedTs;
   private String createdBy;
   private String updatedBy;
-
-  public int getProfAcdmcId() {
+  /**
+   * @return the profAcdmcId
+   */
+  public Long getProfAcdmcId() {
     return profAcdmcId;
   }
-
-  public void setProfAcdmcId(int profAcdmcId) {
+  /**
+   * @param profAcdmcId the profAcdmcId to set
+   */
+  public void setProfAcdmcId(Long profAcdmcId) {
     this.profAcdmcId = profAcdmcId;
   }
-
+  /**
+   * @return the profAcdmcName
+   */
   public String getProfAcdmcName() {
     return profAcdmcName;
   }
-
+  /**
+   * @param profAcdmcName the profAcdmcName to set
+   */
   public void setProfAcdmcName(String profAcdmcName) {
     this.profAcdmcName = profAcdmcName;
   }
-
+  /**
+   * @return the profAcdmcDesc
+   */
   public String getProfAcdmcDesc() {
     return profAcdmcDesc;
   }
-
+  /**
+   * @param profAcdmcDesc the profAcdmcDesc to set
+   */
   public void setProfAcdmcDesc(String profAcdmcDesc) {
     this.profAcdmcDesc = profAcdmcDesc;
   }
-
+  /**
+   * @return the profAcdmcUniversity
+   */
   public String getProfAcdmcUniversity() {
     return profAcdmcUniversity;
   }
-
+  /**
+   * @param profAcdmcUniversity the profAcdmcUniversity to set
+   */
   public void setProfAcdmcUniversity(String profAcdmcUniversity) {
     this.profAcdmcUniversity = profAcdmcUniversity;
   }
-
+  /**
+   * @return the profAcdmcPassYear
+   */
   public String getProfAcdmcPassYear() {
     return profAcdmcPassYear;
   }
-
+  /**
+   * @param profAcdmcPassYear the profAcdmcPassYear to set
+   */
   public void setProfAcdmcPassYear(String profAcdmcPassYear) {
     this.profAcdmcPassYear = profAcdmcPassYear;
   }
-
-  public long getProfId() {
+  /**
+   * @return the profId
+   */
+  public Long getProfId() {
     return profId;
   }
-
-  public void setProfId(long profId) {
+  /**
+   * @param profId the profId to set
+   */
+  public void setProfId(Long profId) {
     this.profId = profId;
   }
-
-
-
-  public int getActiveStatus() {
+  /**
+   * @return the activeStatus
+   */
+  public Integer getActiveStatus() {
     return activeStatus;
   }
-
-  public void setActiveStatus(int activeStatus) {
+  /**
+   * @param activeStatus the activeStatus to set
+   */
+  public void setActiveStatus(Integer activeStatus) {
     this.activeStatus = activeStatus;
   }
-
+  /**
+   * @return the createdTs
+   */
   public Date getCreatedTs() {
     return createdTs;
   }
-
+  /**
+   * @param createdTs the createdTs to set
+   */
   public void setCreatedTs(Date createdTs) {
     this.createdTs = createdTs;
   }
-
+  /**
+   * @return the updatedTs
+   */
   public Date getUpdatedTs() {
     return updatedTs;
   }
-
+  /**
+   * @param updatedTs the updatedTs to set
+   */
   public void setUpdatedTs(Date updatedTs) {
     this.updatedTs = updatedTs;
   }
-
+  /**
+   * @return the createdBy
+   */
   public String getCreatedBy() {
     return createdBy;
   }
-
+  /**
+   * @param createdBy the createdBy to set
+   */
   public void setCreatedBy(String createdBy) {
     this.createdBy = createdBy;
   }
-
+  /**
+   * @return the updatedBy
+   */
   public String getUpdatedBy() {
     return updatedBy;
   }
-
+  /**
+   * @param updatedBy the updatedBy to set
+   */
   public void setUpdatedBy(String updatedBy) {
     this.updatedBy = updatedBy;
   }
-
 }
